Replace clang-tidy with regex on each added line in a patch
Using clang-tidy correctly would require that this check understand a project's entire compile toolchain, so that the correct compile #defines and other flags were used. Since we can't do that, just use a regex to look for added gotos.

I think this should be sufficient for any sanely formatted c/cpp files. This will miss any oneline `if (...) goto symbol;`, but I think that's enough of a minority to not worry about. There are probably other cases I'm forgetting about too. I did worry about the case where `goto` is in a comment, and multiple statements separated by ; on one line. I decided that KISS is the best goal here, and I'm hoping this still covers at least 90% of cases.

This removes the clang-tidy-prebuilt dependency. It also subtly changes the getAnnotations() siganture so that getChangedCFiles() is called outside of it, so that it makes it easier to test getAnnotations(). The new functions are unit-tested, and I started using jest's describe() to group like tests together, since we're getting a good amount.